Customer Analytics Using ML & DL
The project deals with the customer analytics using Python. The main aim of this project is to make customer segmentation using agglomerative  clustering and other methods like K-Mean to come up with ideal segments of customers. And then based on the segmentation, to calculate the demand of elasticity of each customer. Predict customers brand preference based on segments created. In the end using deep learning method to deal with customer retention, i.e. whether a customer is going to purchase a product or not.